Refactor project structure and enhance settings management
- Added pyproject.toml for package management and build system configuration.
- Created src/imap_sorting_hat module structure
- Updated tests to reflect new module structure and import paths.
