+++ .github/workflows/release.yml
@@ -62,7 +62,8 @@ jobs:
       - name: Install dependencies
         run: |
           pip install -U pip
-          pip install -r requirements.txt
+          pip install -r requirements.txt || true
+          pip install -e .
 
       - name: Run tests
         run: pytest tests
