+++ .vscode/settings.json
@@ -7,6 +7,9 @@
     "python.analysis.inlayHints.pytestParameters": false,
     "python.analysis.enablePytestSupport": false,
     "editor.formatOnSave": true,
+    "python.analysis.extraPaths": [
+        "./src"
+    ],
     "python.testing.pytestArgs": [
         "tests"
     ],
