+++ pyproject.toml
@@ -1,0 +1,11 @@
+[project]
+name = "imap-sorting-hat"
+version = "0.0.4"
+requires-python = ">=3.11"
+
+[build-system]
+requires = ["setuptools>=61.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools.packages.find]
+where = ["src"]
