+++ src/imap_sorting_hat/settings.py
